requirements.txt
# Core LangChain libraries
langchain==0.1.20
langgraph==0.0.48
langchain-openai==0.1.6
langchain-groq==0.1.3

# For managing environment variables
python-dotenv==1.0.1

# For graph visualization
networkx==3.2.1
matplotlib==3.8.3
numpy==1.26.4
